refactor: MultiplierEngine extract, shared CRUD adapter, personnel typeahead
