+++ frontend/src/composables/usePersonnelSearch.js
@@ -1,0 +1,21 @@
+import { useApi } from '@/composables/useApi.js'
+
+/**
+ * Shared personnel typeahead — one seam for เกื้อกูล / แตกต่าง / เทียบตำแหน่ง / ทวีคูณ pages.
+ */
+export function usePersonnelSearch() {
+  const api = useApi()
+
+  async function searchPersonnel(query, { limit = 10 } = {}) {
+    const q = (query || '').trim()
+    if (!q) return []
+
+    const params = new URLSearchParams()
+    params.set('search', q)
+    params.set('limit', String(limit))
+    const result = await api.get(`/personnel?${params}`)
+    return result.data || []
+  }
+
+  return { searchPersonnel }
+}
